Reorganize classes, add opencl selfcheck
